changed the way node sends data to base station
added delay between each radio packet to give base station time to pass
it over serial to the computer. Also increased Baud Rate of the serial
interface. This is a temporary solution to avoid data loss in
transmission and a better solution that does not involve delays should
be coded.