Fixed some bugs and introduced delay between subsequent messages
fixed bugs related to writing data to FRAM as it is collected by the ADC
and also fixed a bug where the sample count did not reset from one call
of a sample function to the next.

Also, had to introduce a 100ms delay between subsequent messages in
order to give the base station time to pass data via USB. Without the
delay, messages would be missed while the base station is busy writing
previous data to USB. The RAM on the base station is also not large
enough to buffer all of the data that can potentially be transmitted
from one mote. This is a crude solution and needs to be changed if
possible.